stress-tsearch: assign data the correctly typed values
Cast the computed random values to the same type as data

+++ stress-tsearch.c
@@ -92,7 +92,7 @@ static int stress_tsearch(const stress_args_t *args)
 
 		/* Step #1, populate tree */
 		for (i = 0; i < n; i++) {
-			data[i] = ((stress_mwc32() & 0xfff) << 20) ^ i;
+			data[i] = (int32_t)(((stress_mwc32() & 0xfff) << 20) ^ i);
 			if (tsearch(&data[i], &root, cmp) == NULL) {
 				size_t j;
